convert emc values to doubles for proper formatting

+++ src/main/java/cool/furry/mc/forge/projectexpansion/gui/EMCDisplay.java
@@ -64,8 +64,8 @@ public class EMCDisplay {
     @SubscribeEvent
     public static void onRenderGUI(RenderGameOverlayEvent.Text event) {
         if(!Config.emcDisplay.get()) return;
-        String str = EMCFormat.INSTANCE.format(emc);
-        if(!change.equals(BigInteger.ZERO)) str += " " + (change.compareTo(BigInteger.ZERO) > 0 ? (TextFormatting.GREEN + "+") : (TextFormatting.RED + "-")) + EMCFormat.INSTANCE.format(change) + "/s";
+        String str = EMCFormat.INSTANCE.format(emc.doubleValue());
+        if(!change.equals(BigInteger.ZERO)) str += " " + (change.compareTo(BigInteger.ZERO) > 0 ? (TextFormatting.GREEN + "+") : (TextFormatting.RED + "-")) + EMCFormat.INSTANCE.format(change.doubleValue()) + "/s";
         event.getLeft().add(String.format("EMC: %s", str));
     }
 }
